Refactor utils to new folder
