+++ TPI/juegoDeLaVida/es.cpp
@@ -1,6 +1,6 @@
 #include "es.h"
 #include "definiciones.h"
-#include "toroidUtils.h"
+#include "utils/toroidUtils.h"
 #include <iostream>
 #include <fstream>
