+++ TPI/juegoDeLaVida/utils/toroidUtils.cpp
